Revert "Added thr Problem Statement Component with redux store and redux saga"
This reverts commit c32c0d1fb56008d35f8f65d121858453fc14dc18.
