--- src/apis/problemStatementAPI.js
@@ -1,7 +1,0 @@
-const axios = require('axios').default;
-
-const getStatement = () => {
-   return axios.get('http://localhost:3000/statementsList');
-}
-
-export default getStatement;
