+++ src/apis/rulesAPI.js
@@ -1,7 +1,7 @@
 const axios = require('axios').default;
 
 const getRules = () => {
-   return axios.get('  http://localhost:3002/rulesList');
+   return axios.get('http://localhost:3000/rulesList');
 }
 
 export default getRules;
