--- src/components/ProblemsComponent/problemStyle.css
@@ -1,4 +1,0 @@
-.problemBody {
-    background-color: #292727;
-    height: 100vh;
-  }
