+++ src/components/ProfileComponent/index.css
@@ -2,7 +2,3 @@
     background-color: #218838;
     height: 100vh;
   }
-.grey {
-  background-color: #292727;
-  height: 100vh;
-}
