+++ src/components/UserProfileComponent/index.jsx
@@ -1,7 +1,6 @@
-
 const UserProfileComponent = () => {
     return(
-        <h4 className = "text-white text-center">UserProfileComponent</h4>
+        <h1>UserProfileComponent</h1>
     )
 }
