+++ src/constants/actionConstants.js
@@ -1,10 +1,7 @@
-export const PROBLEM_STATEMENT_REDUCER = {
-    SET_STATEMENT: 'SET_STATEMENT',
-    STATEMENT_REQUEST: 'STATEMENT_REQUEST'
-}
-
+// export const ACTION_CONSTANT = "ACTION_CONSTANTS";
 const RULES_REDUCER = {
     SET_RULES: 'SET_RULES',
     RULES_REQUEST: 'RULES_REQUEST'
 }
 export default RULES_REDUCER;
+
