+++ src/store.js
@@ -1,12 +1,12 @@
-import { createStore, applyMiddleware, combineReducers} from "redux";
+import { createStore, applyMiddleware} from "redux";
 import createSagaMiddleware from "redux-saga";
 
-import ProblemStatementReducer from "reducers/ProblemStatementReducer";
+// import rootReducer from "reducers/rootReducer";
 import RulesReducer from "reducers/RulesReducer";
 import rootSagas from "sagas/rootSagas";
 
 const sagaMiddleware = createSagaMiddleware();
 
-export const store = createStore(combineReducers({RulesReducer,ProblemStatementReducer}),applyMiddleware(sagaMiddleware));
+export const store = createStore(RulesReducer,applyMiddleware(sagaMiddleware));
 
 sagaMiddleware.run(rootSagas);
